requirements.txt
# Requirements for PhySR
# Install via conda : conda install --file requirements.txt

pytorch == 1.11.0
sympy == 1.10.1
numpy
matplotlib
pandas
scikit-learn
jupyterlab
pip